--- a/node/impl/full/eth.go
+++ b/node/impl/full/eth.go
@@ -980,14 +980,14 @@ func (a *EthModule) applyMessage(ctx context.Context, msg *types.Message, tsk ty
 		return nil, xerrors.Errorf("cannot get tipset: %w", err)
 	}
 
-	applyTsMessages := true
-	if os.Getenv("LOTUS_SKIP_APPLY_TS_MESSAGE_CALL_WITH_GAS") == "1" {
-		applyTsMessages = false
-	}
-
 	// Try calling until we find a height with no migration.
 	for {
-		res, err = a.StateManager.CallWithGas(ctx, msg, []types.ChainMsg{}, ts, applyTsMessages)
+		st, _, err := a.StateManager.TipSetState(ctx, ts)
+		if err != nil {
+			return nil, xerrors.Errorf("cannot get tipset state: %w", err)
+		}
+
+		res, err = a.StateManager.ApplyOnStateWithGas(ctx, st, msg, ts)
 		if err != stmgr.ErrExpensiveFork {
 			break
 		}
@@ -997,7 +997,7 @@ func (a *EthModule) applyMessage(ctx context.Context, msg *types.Message, tsk ty
 		}
 	}
 	if err != nil {
-		return nil, xerrors.Errorf("CallWithGas failed: %w", err)
+		return nil, xerrors.Errorf("ApplyWithGasOnState failed: %w", err)
 	}
 	if res.MsgRct.ExitCode.IsError() {
 		reason := parseEthRevert(res.MsgRct.Return)