@aarshkshah1992 2024-06-11
@raulk Vulcanise is also running into some of the rough edges related to the events. So we are starting the work with fixing one of the issues they're running into i.e. async indexing of events leading to races where client discovers a block but does not get any events for them because they're yet to be indexed. 

However, please note that we don't have dedicated epic to fix all of the issues you and @rvagg have enlisted. If doing all of those is important, would be great to do a prioritsation epic for this workstream with @jennijuju and @rvagg.